+++ src/rtweekend.zig
@@ -1,9 +1,9 @@
 const std = @import("std");
 
-pub fn degrees_to_radians(degrees: f32) f32 {
+pub fn degrees_to_radians(degrees: f64) f64 {
     return degrees * std.math.pi / 180.0;
 }
 
-pub fn rand_range(min: f32, max: f32) f32 {
-    return min + std.rand.float(std.Random.Xoshiro256, f32) * (max - min);
+pub fn rand_range(min: f64, max: f64) f64 {
+    return min + std.rand.float(std.Random.Xoshiro256, f64) * (max - min);
 }
